Passage des ID en Long, prix en BigDecimal, ajout du stock et de la documentation

+++ src/main/java/org/romain/demo2/dao/ProductDao.java
@@ -7,7 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository // Met en place le systeme de query method et ...
-public interface ProductDao extends JpaRepository<Product, Integer> {
-
+public interface ProductDao extends JpaRepository<Product, Long> {
 
 }
